+++ events/guild_messages/messageCreate.js
@@ -1,6 +1,11 @@
 const { MessageEmbed, Permissions } = require('discord.js')
+
 const langFr = require('../../languages/fr/events/messageCreate.json')
 const langEn = require('../../languages/en/events/messageCreate.json')
+
+const talkedRecently = new Set()
+const { REST } = require('@discordjs/rest')
+const { Routes } = require('discord-api-types/v9')
 
 module.exports = {
   name: 'messageCreate',
@@ -12,16 +17,34 @@ module.exports = {
     if (message.author.bot) return
     let guildSettings = await client.getGuild(message.guild)
 
+    if (!message.member.permissions.has([Permissions.FLAGS.ADMINISTRATOR])) {
+      if (talkedRecently.has(message.author.id)) {
+        message.delete()
+        message.channel.send({ content: `<@${message.author.id}> - ${lang}` })
+      } else {
+        // the user can type the command ... your command code goes here :)
+
+        // Adds the user to the set so that they can't talk for a minute
+        talkedRecently.add(message.author.id)
+        setTimeout(() => {
+          // Removes the user from the set after a minute
+          talkedRecently.delete(message.author.id)
+        }, 60000)
+      }
+    }
+
     const guildBadWords = guild.badWords
 
     if (!message.member.permissions.has([Permissions.FLAGS.ADMINISTRATOR])) {
-      for (const word of guildBadWords) {
-        if (message.content.includes(word)) {
-          message.channel.send(`<@${message.author.id}>\n${lang.autoModErreur1} \`${word}\` ${lang.autoModErreur2}`)
-          setTimeout(() => {
-            message.delete()
-          }, 5000)
-          return
+      if (guild.autoModActive) {
+        for (const word of guildBadWords) {
+          if (message.content.includes(word)) {
+            message.channel.send(`<@${message.author.id}>\n${lang.autoModErreur1} \`${word}\` ${lang.autoModErreur2}`)
+            setTimeout(() => {
+              message.delete()
+            }, 5000)
+            return
+          }
         }
       }
     }
@@ -32,7 +55,7 @@ module.exports = {
     }
 
     if (message.author.id === message.guild.ownerId) {
-      if (message.content === '!slash') {
+      if (message.content === '!slash on') {
         const guildObject = await client.guilds.cache.get(message.guild.id)
         await guildObject.commands.set(client.commands.map(cmd => cmd))
         guild.save().then(async () => {
@@ -41,8 +64,28 @@ module.exports = {
             .setColor('GREEN')
             .setFooter({ text: `${lang.footer} ${message.author.tag}`, avatarURL: `${message.author.displayAvatarURL(true)}` })
           await message.delete()
-          message.channel.send({ embeds: [embed] })
+          await message.channel.send({ embeds: [embed] })
         })
+      } else if (message.content === '!slash off') {
+        const reply = await message.reply('Veuillez patienter. L\'oppération peut prendre un certain temps')
+        const token = process.env.YMULE
+        const clientId = process.env.CLIENT_ID
+        const guildId = message.guild.id
+
+        const rest = new REST({ version: '9' }).setToken(token)
+        rest.get(Routes.applicationGuildCommands(clientId, guildId))
+          .then(data => {
+            const promises = []
+            for (const command of data) {
+              const deleteUrl = `${Routes.applicationGuildCommands(clientId, guildId)}/${command.id}`
+              promises.push(rest.delete(deleteUrl))
+            }
+            return Promise.all(promises)
+          }).then(() => {
+            reply.delete()
+            message.delete()
+            message.channel.send('Les commandes ont correctement été désactivés !')
+          })
       }
     }
   }
